goldenraspberry/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
pandas==2.2.2